+++ include/pomerol/GFContainer.hpp
@@ -41,6 +41,10 @@ namespace Pomerol {
 /// \f[
 ///  G_{ij}(i\omega_n) = -\int_0^\beta d\tau e^{i\omega_n\tau} Tr[\mathcal{T}_\tau \hat\rho c_i(\tau) c_j^\dagger(0)].
 /// \f]
+/// or of an anomalous Green's function
+/// \f[
+///  F_{ij}(i\omega_n) = -\int_0^\beta d\tau e^{i\omega_n\tau} Tr[\mathcal{T}_\tau \hat\rho c_i(\tau) c_j(0)].
+/// \f]
 class GFContainer : public IndexContainer2<GreensFunction, GFContainer>, public Thermal {
 
 public:
@@ -56,31 +60,34 @@ public:
     /// \param[in] H The Hamiltonian.
     /// \param[in] DM Many-body density matrix \f$\hat\rho\f$.
     /// \param[in] Ops A set of creation/annihilation operators \f$c^\dagger_j\f$/\f$c_i\f$.
+    /// \param[in] Anomalous Store anomalous Green's functions.
     template <typename... IndexTypes>
     GFContainer(IndexClassification<IndexTypes...> const& IndexInfo,
                 StatesClassification const& S,
                 Hamiltonian const& H,
                 DensityMatrix const& DM,
-                FieldOperatorContainer const& Ops)
+                FieldOperatorContainer const& Ops,
+                bool Anomalous = false)
         : IndexContainer2<GreensFunction, GFContainer>(*this, IndexInfo),
           Thermal(DM),
           S(S),
           H(H),
           DM(DM),
-          Operators(Ops) {}
+          Operators(Ops),
+          Anomalous(Anomalous) {}
 
-    /// Prepare a set of matrix elements \f$G_{ij}\f$.
-    /// \param[in] Indices Set of index combinations of the elements \f$G_{ij}\f$ to be prepared.
+    /// Prepare a set of matrix elements \f$G_{ij}\f$ or \f$F_{ij}\f$.
+    /// \param[in] Indices Set of index combinations of the elements \f$G_{ij}\f$/\f$F_{ij}\f$ to be prepared.
     ///            An empty set results in creation of elements for all possible index combinations.
     void prepareAll(std::set<IndexCombination2> const& Indices = {});
-    /// Compute all prepared matrix elements \f$G_{ij}\f$.
+    /// Compute all prepared matrix elements \f$G_{ij}\f$/\f$F_{ij}\f$.
     /// \pre \ref prepareAll() has been called.
     void computeAll();
 
 protected:
     friend class IndexContainer2<GreensFunction, GFContainer>;
 
-    /// Create a single matrix element \f$G_{ij}\f$.
+    /// Create a single matrix element \f$G_{ij}\f$/\f$F_{ij}\f$.
     /// \param[in] Indices Index combination \f$(i,j)\f$.
     std::shared_ptr<GreensFunction> createElement(IndexCombination2 const& Indices) const;
 
@@ -93,6 +100,8 @@ protected:
     DensityMatrix const& DM;
     /// A set of creation/annihilation operators \f$c^\dagger_j\f$/\f$c_i\f$.
     FieldOperatorContainer const& Operators;
+    /// Store anomalous Green's functions.
+    bool const Anomalous;
 };
 
 ///@}