+++ CHANGELOG.md
@@ -12,6 +12,9 @@ All notable changes to this project will be documented in this file.
   achieved by passing two `AnnihilationOperator` objects to the constructor.
   A unit test for this new functionality has been added.
 
+- `GFContainer` can now store anomalous Green's functions (an extra boolean
+  argument has been added to its constructor to control this new feature).
+
 - New class `QuarticOperator`. In full analogy with `QuadraticOperator`, it is
   meant to be primarily used with `EnsembleAverage` to compute e.g. average
   double occupancy.